feat: docker bake tracking version tag
For a given Slurm version (e.g. 24.11.x), we want:
24.11-$FLAVOR -> 24.11.x-$FLAVOR

This commit makes the docker bake responsible for all tags. The CI will
still pull the built image layers, but docker bake will be invoked again
to re-tag the images and push them (technically will build if layers are
different).