fix: run sackd process in login image as root
It is preferrable to run the sackd process as non-root. However, the
login image requires to be run as root for sshd and other processes to
function as intended.

Kubernetes lacks the ability to control select volume mounts as being
mounted under a user context, due to the single process principle of
containers. Hence, we cannot mix root and non-root volumes. Host SSH
keys must be strictly 0o600 but Slurm secrets can be 0o640, therefore we
are limited by sshd than sackd.